backend-lummia-platform/requirements.txt
flask==3.1.1
flask-cors==5.0.1
flask-socketio==5.5.1
psycopg2-binary==2.9.10
pymongo==4.12.1
bcrypt==4.2.1
PyJWT==2.10.1
python-dotenv==1.1.0
google-genai==1.14.0
gunicorn==23.0.0
gevent==24.11.1
gevent-websocket==0.10.1
